Move Node component to containers
