+++ src/containers/index.js
@@ -9,6 +9,7 @@ export { default as PromptSwiper } from './PromptSwiper';
 export { default as Form } from './Form';
 export { default as NodePanels } from './NodePanels';
 export { default as Field } from './Field';
+export { default as Node } from './Node';
 export { default as NodeForm } from './NodeForm';
 export { default as FormWizard } from './FormWizard';
 export { default as ConcentricCircles } from './ConcentricCircles';
